WIP move adaptive_experiment.py
