+++ examples/synthetic_graph_evolution/graph_search.py
@@ -6,6 +6,7 @@ from examples.synthetic_graph_evolution.experiment import run_experiments
 from examples.synthetic_graph_evolution.generators import generate_labeled_graph
 from golem.core.adapter.nx_adapter import BaseNetworkxAdapter
 from golem.core.dag.verification_rules import has_no_self_cycled_nodes
+from golem.core.optimisers.adaptive.operatoragent import MutationAgentTypeEnum
 from golem.core.optimisers.genetic.gp_optimizer import EvoGraphOptimizer
 from golem.core.optimisers.genetic.gp_params import GPAlgorithmParameters
 from golem.core.optimisers.genetic.operators.base_mutations import MutationTypesEnum
@@ -36,6 +37,7 @@ def graph_search_setup(target_graph: nx.DiGraph,
         history_dir=None,
     )
     gp_params = GPAlgorithmParameters(
+        adaptive_mutation_type=MutationAgentTypeEnum.random,
         pop_size=21,
         multi_objective=False,
         genetic_scheme_type=GeneticSchemeTypesEnum.generational,
